Updating the version_update function in models.py
Added the beginnings of a version update function that doesn't search for AAAA's. This one is still called version_update_new, once it's fully ready to go I'll change it to version_update (and either rename the old function or delete it).

This is still a WIP (only updates version_id tag and needs more testing), but I'm committing this early version now, while it works, because I'm definitely about to break it