More version_update stuff
The new function now puts the version number inside the correct tag (<information_model_version>, not <version_id>, oops) .

It also now updates the version numbers in the schema location within <Product_Observational> as well.

Right now, this doesn't do anything to any additional discipline dictionaries that a label might have- those will need extra steps to update because there's a second version number in addition to the regular pds4 information model number. Messing with that can wait until we're working on more support for discipline dictionaries in general, but I'm putting it on the record.