integrate into the djangoverse, plus saving to db
re: integration with django: this script now runs in that environment and accesses the db correctly. I'll eventually make this whole thing a command that runs from manage.py.

re: saving new context products to the db: I now know how to do it and have the test code in here, but might not do much with it until we've made some more decisions about changing up the context section of the website. What this code needs to put into what tables here is connected to how the user does it on the website, and how the user does context on the website is (i think / hope) going to be changed and improved fairly soon, so,